For just the simple tests we have, this changed the integration test duration from 236 secs to 155 secs on my laptop.