@ErikDeSmedt 2024-08-14
Are we even running the same tests. I have the following timings

- master: 23s
- round-timgs: 26s