@stevenroose 2024-08-14
I'm counting the entire invocation, without compilation. `time just test-integration`. My laptop has been laggy lately, but I can't imagine it's a 10x slower than yours :sweat_smile: