@stevenroose 2024-08-14
This might give a glimps
```

[16:33:26.201  INFO] Starting aspd-1
[16:33:26.201 TRACE] Preparing aspd-1
test check_aspd_version has been running for over 60 seconds
test fund_asp has been running for over 60 seconds
test check_aspd_version ... ok
[16:34:57.122 TRACE] aspd-1: Trying to spawn "cargo" "run" "--package" "bark-aspd" "--" "--datadir" "/home/steven/code/rust/bark/test/aspd/fund_aspd/aspd-1" "start"
[16:34:59.690  INFO] Started aspd-1
```

The fact that 2 tests are trying to setup an aspd at the same time might not be ideal.