@stevenroose 2024-08-14
Running again, I get 337s and 195s. Seems to be wildly unreliable, but I might get some benefit from understanding why that setting up aspd is taking so long. It seems to be the bulk of it.