@stevenroose 2024-08-15
Even not doing multiple tests:
`time just test-integration onboard_bark` takes 11s
`time just test-integration multiple_round_payments` takes 88s.

It's hard to use trace logging to find out what's happening under the hood because the logs emissions themselves are subject to the tokio runtime scheduler, I think.

Generally, we might benefit from running integration tests with `--jobs 1` to avoid any races between the tests and to make logs more readable. For me it's not hurting runtime, it's actually slightly better on the times I tried.