@stevenroose 2024-08-15
Added a few more things. Could you check if this is hindering you at all?