@ErikDeSmedt 2024-08-19
I

> The fact that 2 tests are trying to setup an aspd at the same time might not be ideal.

I think I understand what the problem with slow tests might be.
By default, we run `cargo run --package aspd --` instead of `aspd`.
It appears that on every command you compile `aspd` over and over again.
On my machine the result is cached and the second invocation is immediate.

Could you try to do 

```
cargo build
export ASPD_EXEC=$PWD/target/debug/aspd
export BARK_EXEC=$PWD/target/debug/bark
time just test-integration
```

This should ensure we don't do the compilation-step.