@stevenroose 2024-08-19
I don't think I'm compiling it because I see the compilation logs only at the beginning. But let me try that.