@stevenroose 2024-08-19
Oh that's a lot faster.

```
just test-integration  28.11s user 6.76s system 112% cpu 31.130 total
```
on master and 


```
just test-integration  27.55s user 5.96s system 127% cpu 26.245 total
```

on this branch.

Yeah then the diff is negligible. 

Then we might not need the last commit. I think explicitly setting round times is probably a good idea anyway.

I get this if I remove the last commit and parallelize again:
```
cargo test --package ark-testing  23.01s user 4.90s system 132% cpu 21.082 total
```